+++ webapp/cognitive-coach/src/components/ConfigurePopup/ConfigurePopup.css
