+++ webapp/cognitive-coach/src/components/ConfigurePopup/ConfigurePopup.tsx
