+++ webapp/cognitive-coach/src/components/ConfigurePopup/ScreenRecordingPreview.tsx
