+++ webapp/cognitive-coach/src/components/ConfigurePopup/WebcamPreview.tsx
