+++ webapp/cognitive-coach/src/components/CurrentSession/CurrentSession.css
