+++ webapp/cognitive-coach/src/pages/SessionDetail/SessionDetail.css
